% $Id$

\name{DeMixSampleData-package}
\alias{DeMixSampleData-package}
\docType{package}
\title{\Sexpr[results=text]{packageDescription("DeMixSampleData")$Title}}
\description{
  \Sexpr[results=text]{packageDescription("DeMixSampleData")$Description}
}
\details{
  \tabular{ll}{
    Package: \tab \Sexpr[results=text]{packageDescription("DeMixSampleData")$Package}\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("DeMixSampleData")$Version}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("DeMixSampleData")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("DeMixSampleData")$License}\cr
  }

  For a complete list of functions, use \code{library(help="DeMixSampleData")}.\cr
  For a high-level summary of the changes for each revision, use\cr
    \code{file.show(system.file("NEWS", package="DeMixSampleData"))}.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

